requirements.txt
asgiref==3.7.2
Flask==3.0.0
python-telegram-bot==20.6
uvicorn==0.23.2
python-dotenv